+++ src/main/kotlin/io/github/firstred/iptvproxy/dtos/xtream/XtreamCategoryIdServer.kt
@@ -3,8 +3,8 @@ package io.github.firstred.iptvproxy.dtos.xtream
 import io.github.firstred.iptvproxy.enums.IptvChannelType
 
 data class XtreamCategoryIdServer(
-    val id: Long,
-    val externalId: Long,
+    val id: UInt,
+    val externalId: UInt,
     val server: String,
     val type: IptvChannelType,
 )
